Created MidiPerformance class and updated AudioGenerator (#174)
* created script to synthesize and play notes in real time

* added fluidsynth download for windows to setup.py

* updated .gitignore

* synth is working but tempo is fixed and audio does not sound like a cello

* created synthesizer version using fluidsynth

* updated synthesizer to allow for tempo changes

* created MidiPerfrmance class

* renamed synthesizer.py to midi_performance.py

* integrated MidiPerformance with main.py

* added midi files

* AudioGenerator no longer changes original midi file

* added more midi files and updated readme

* updated README

* uncommented parts of app.py

* reverted more changes to app.py

+++ backend/src/score_follower.py
@@ -91,7 +91,7 @@ class ScoreFollower:
         # Return a chroma feature for the audio
         return self.chroma_maker.insert(audio)
 
-    def step(self, frames: np.ndarray) -> int:
+    def step(self, frames: np.ndarray) -> float:
         """Calculate next step in the alignment path between the microphone and reference audio """
 
         # Generate chroma feature
